Add admin route and empty container for admin dashboard.
Not sure if piping the admin dashboard through the pages controller is the right thing todo, but it's the easiest at the moment.

Added an extra admin layout since I don't wont to pollute the main layout with it. Maybe we should devide the dashboard and the normal site even further (seperate directories).