Add initial react setup for admin interface.
Setup used:
- react
- redux
- react-router(-redux)
- redux-thunk

This gives the initial structure and libraries for the frontend development. React is used as the main library for dom rendering. Redux is used to manage the app state. To handle client-side routing I introduces react-router.

Further I added react-router-redux to sync the router with the app state. This will probably be handy in the future and it's easiest to introduce it right from the get go. redux-thunk was added to make async actions a smooth experience.