Add sidenav for admin interface
Simple sidenav that is expanded on screens wider 768pxs by default and collapsed otherwise. I made it 250px wide for now, but this might be to large.

Implementation nitpicks:
I think it's not ideal that the 'toggled' class on the shell-wrappper div means the exact opposite on small vs large screens. This is not a problem at the moment (except maybe that the sidenav vanishes if you wider your window), but it will be a hassle if we try to implement auto hiding the navbar on item click only on small screens.