Add User model
Tests are just a valid and an empty map. Needs to be improved.

Since there probably should be a default admin user seeded we maybe need to add usernames later, but I would like to avoid them if possible and just use email and password. First and last names were added in case users want to display authorship.