Add session logic to frontend
Note that this implements a DELETE_SESSION action but does'nt add a way to trigger it in the UI. This is because I want to add that to the sidenav and it needs refactoring first.

This commit stores the JWT in localStorage. This has some bad implections for things like XSS vulnerabilities. It should probably be replaced by cookies.