Switch from localStorage to Cookies for JWT storage.
This will probably be the better choice because it clears the session on browser close. I think this the expected behaviour for most users.

Maybe we should ad a "Remember me" option that stores the cookie for the time it is valid.

I also fixed a bug with the polyfilling of fetch. Not sure if it works in IE now, since we don't polyfill promises.