Switch to react bootstrap (#7)
* Switch from bootstrap 4 to 3 (Yeti theme)

Rip out bootstrap 4 from the build and use bootstrap 3 with yeti theme instead. Currently it gets pullled from MaxCDN instead of building it for the project. Mainly because setting up webpack with bootstrap 3 is tedious and error prone.

This introduces a lot of UI bugs because of the missing bootstrap 4 classes but is needed for the switch to react-bootstrap.

* Switch to react-bootstrap