Fix indentation of chained method calls
I had a different versions then the one that got installed following the package.json of one of the linter packages that specified how to indent the chained method calls. That broke the build.

I thought about shrinkwrapping the linter packages. But I for now I just re-npm-installed and fixed the problems. Is this is a reoccuring problem I might do it though.