Add integration tests for login.
This adds integration tests for login. They introduce quite a bit of time overhead to the test suite, as you would expect, but they are the only good way to test if the cookies are working correctly.

I also added polyfills for `Object.assign` and Promises to `admin.html.eex`. I'm not sure if I want them in the project as I'm not sure if we should support browsers that old, but we need them for PhantomJS anyway so I but them in there for now.

If we decide not to support those browsers it would be great to not load the polyfills outside of PhantomJS/Hound. I opended [this issue](https://github.com/HashNuke/hound/issues/120) on Hound to maybe get some suggestions how to do that.