Fix PhantomJS download in CI
* Change PhantomJS source back to my server

For some reason the official mirror is now failing with wget. It's the
same on my machine.

I switched back to my mirror, that failed last time because of the
cert which seemed to be fine.

* Update wget in CI

The cert related fail was based on an outdated version of wget on travis.
Not sure what the bug is updating it solves the error.

--

That all is kind of a mess. We are updating know wget and PhantomJS every
time because they are outdated on the base image. I hope this gets fixed soon.

Keep watching https://github.com/travis-ci/travis-ci/issues/3225 !